Add group homomorphisms to iset.mm (#4827)
* copy GrpHom from set.mm to iset.mm

This is the syntax and df-ghm

* copy reldmghm from set.mm to iset.mm

* Add isghm to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy isghm3 from set.mm to iset.mm

* copy ghmgrp1 from set.mm to iset.mm

* copy ghmgrp2 from set.mm to iset.mm

* copy ghmf from set.mm to iset.mm

* copy ghmlin from set.mm to iset.mm

* copy ghmid from set.mm to iset.mm

* copy ghminv from set.mm to iset.mm

* copy ghmsub from set.mm to iset.mm

* copy isghmd from set.mm to iset.mm

* copy ghmmhm from set.mm to iset.mm

* copy ghmmhmb from set.mm to iset.mm

* copy ghmmulg from set.mm to iset.mm

Also, fix the comment, including in set.mm, to say group homomorphism.

* Add ghmrn to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing to change not
empty to inhabited but is otherwise the set.mm proof.

* copy 0ghm from set.mm to iset.mm

* copy idghm from set.mm to iset.mm

* Add resghm to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy submmnd from set.mm to iset.mm

* Add submbas to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is basically
the set.mm proof.

* copy subm0 from set.mm to iset.mm

* Add subsubm to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add resmhm to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add resmhm2 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add resmhm2b to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* copy resghm2 from set.mm to iset.mm

* copy resghm2b from set.mm to iset.mm

* copy ghmghmrn from set.mm to iset.mm

* copy ghmco from set.mm to iset.mm

* copy ghmima from set.mm to iset.mm

* Add ghmpreima to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* copy ghmeql from set.mm to iset.mm

* copy ghmnsgima from set.mm to iset.mm

* copy ghmnsgpreima from set.mm to iset.mm

* copy ghmker from set.mm to iset.mm

* copy ghmeqker from set.mm to iset.mm

* Add pwsdiagghm to mmil.html

* copy f1ghm0to0 from set.mm to iset.mm

* copy ghmf1 from set.mm to iset.mm

* Add kerf1ghm to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in places but is
basically the set.mm proof.

* copy ghmf1o from set.mm to iset.mm

* Add conjghm to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in places but is
basically the set.mm proof.

* copy conjsubg from set.mm to iset.mm

* copy conjsubgen from set.mm to iset.mm

* Add conjnmz to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in places but is
basically the set.mm proof.

Also, minimize conjnmz in set.mm.

* copy grpassd from set.mm to iset.mm

* Add conjnmzb to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

* copy conjnsg from set.mm to iset.mm

* copy eqg0el from set.mm to iset.mm

* Add quselbasg to iset.mm

This is quselbas from set.mm with an additional set existence condition.
The proof is similar to the set.mm proof.

* Add quseccl0g to iset.mm

This is quseccl0 from set.mm with an added set existence condition.  The
proof is based on the set.mm proof.

* copy qusgrp from set.mm to iset.mm

* Add quseccl to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* copy qusadd from set.mm to iset.mm

* Add qusghm to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy ghmpropd from set.mm to iset.mm