Expand ring homomorphisms in iset.mm (#4838)
* Rename df-rnghom and df-rngiso in iset.mm

That is, rename df-rnghom to df-rhm and df-rngiso to df-rim .
This is to match set.mm.

* copy sbbidv from set.mm to iset.mm

* Add eqabdv to iset.mm

Copied from set.mm with the only change being to the comment, to remove
reference to an axiom which iset.mm doesn't have (at least not under the
same name).

* Add dfrhm2 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* copy rhmrcl1 from set.mm to iset.mm

* copy rhmrcl2 from set.mm to iset.mm

* Add mhmex to iset.mm

Set existence theorem for MndHom .

* Add ghmex to iset.mm

This is a set existence theorem for GrpHom .

* Add isrhm to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* copy rhmmhm from set.mm to iset.mm

* add rhmisrnghm to mmil.html

* copy rimrcl from set.mm to iset.mm

* copy elovmpod from set.mm to iset.mm

* Add rhmex to iset.mm

set existence theorem for ring homomorphism.

* Add isrim0 to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* copy rhmghm from set.mm to iset.mm

* copy rhmf from set.mm to iset.mm

* Add rhmmul to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add isrhm2d to iset.mm

Stated as in set.mm.  The proof needs intuitionizing for set existence
conditions but is basically the set.mm proof.

* copy isrhmd from set.mm to iset.mm

* Add rhm1 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing for set existence
conditions but is otherwise the set.mm proof.

* Add idrhm to mmil.html

* Add rhmf1o to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionization
for set existence but is basically the set.mm proof.

* copy isrim from set.mm to iset.mm

* copy rimf1o from set.mm to iset.mm

* Add rimrhm to iset.mm

Copied from set.mm with the only change being to the comment, to remove
a reference to a theorem iset.mm doesn't have.

* add rimgim and rimisrngim to mmil.html

* Add rhmfn to iset.mm

Stated as in set.mm.  The proof needs intuitionizing but is based on the
set.mm proof.

* Add rhmval to iset.mm

Stated as in set.mm.  The proof needs a little intuitionizing but is
basically the set.mm proof.

* copy rhmco from set.mm to iset.mm

* add pwsco1rhm to mmil.html

* add pwsco2rhm to mmil.html

* Add brric to mmil.html

Also brrici , brric2 , and ricgic .

* Add rhmdvdsr to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add rhmopp to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in quite a few
places but is basically the set.mm proof.

* Add elrhmunit to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in various places
but is basically the set.mm proof.

* Add rhmunitinv to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in various places
but is basically the set.mm proof.