add znval to iset.mm (#4841)
* Add mulgghm2 to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add mulgrhm to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in several places
but is basically the set.mm proof.

* copy mulgrhm2 from set.mm to iset.mm

* Add zrhvalg to iset.mm

This is zrhval from set.mm with a set existence condition added.  The
proof is based on a portion of the set.mm proof.

* Add zrhval2 to iset.mm

Stated as in set.mm.  The proof needs a small amount of intuitionizing
but is basically the set.mm proof.

* Add zrhmulg to iset.mm

Stated as in set.mm.  The proof needs some intuitionizing but is
basically the set.mm proof.

* Add zrhrhmb to iset.mm

Stated as in set.mm.  The proof needs a little bit of intuitionizing but
is basically the set.mm proof.

Also add zrhex to iset.mm which is a set existence theorem for ZRHom .

* copy zrhrhm from set.mm to iset.mm

* copy zrh1 from set.mm to iset.mm

* copy zrh0 from set.mm to iset.mm

* revise mmil.html entry for zrhpropd

* copy zncrng2 from set.mm to iset.mm

The only change is to the comment, to remove a reference to some syntax
that iset.mm doesn't have.

* Add znval to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in a variety of
places but is based on the set.mm proof.

* Fix comments in set.mm and iset.mm

For mulgghm2 add a sentence saying that .1. here can be any element.

For zrhrhmb change Z to ZZ .

For zncrng2 remove a bunch of text which doesn't apply here very
well and replace it with a comment describing this theorem.