Add 4sqlemsdc to iset.mm
This proves decidability of the set containing numbers which are the sum
of four squares (this set will be used in many of the remaining 4sq lemmas).
The proof involves showing that (for a particular possible element) there
are only a finite number of possible numbers in order to determine whether
the candidate numbers if the sum of four squares.

Includes 4sqexercise1 and 4sqexercise2 which I used in figuring out how
to prove this lemma.  They are not strictly needed any more but are kept
because they may help others in understanding the 4sqlemsdc proof.