prove two 4sq lemmas in iset.mm (#4849)
* Expand 2sq content in mmil.html

In the section on Metamath 100 theorems, update to reflect what we have
done so far.

* Revise lgseisen entry in mmil.html

iset.mm now has Z/nZ and ZRHom

* Add 4sqlemafi to iset.mm

That A is finite is proved by ssfi in set.mm and can be proved here,
although not as easily.

* Add 4sqlemffi to iset.mm

In set.mm, this is part of the proof of 4sqlem11 and is proved using
ssfi .  It is provable here too, but not as easily.

* Add infidc to iset.mm

This is not that different form ssfidc but might as well have this
version too.

* Add 4sqleminfi to iset.mm

In set.mm this is part of the 4sqlem11 proof and follows from ssfi .  We
are also able to prove it, but it is more involved.

* Rename addid1 to addrid in iset.mm

This is to follow a previous rename in set.mm.

* Rename addid2d to addlidd in iset.mm

This is to follow a previous rename in set.mm.

* Add 4sqlem11 to iset.mm

Stated as in set.mm.  Although the proof needs intuitionizing in quite a
few places, many steps in the proof are taken from the set.mm proof
without change.

* Add 4sqlem12 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in a variety of
places but most of the proof is unchanged from the set.mm proof.