Intuitionize 4sqlem13 (#4869)
* copy zzlesq from set.mm to iset.mm

Also, copy the nnlesq comment.

* Add 4sqlemsdc to iset.mm

This proves decidability of the set containing numbers which are the sum
of four squares (this set will be used in many of the remaining 4sq lemmas).
The proof involves showing that (for a particular possible element) there
are only a finite number of possible numbers in order to determine whether
the candidate numbers if the sum of four squares.

Includes 4sqexercise1 and 4sqexercise2 which I used in figuring out how
to prove this lemma.  They are not strictly needed any more but are kept
because they may help others in understanding the 4sqlemsdc proof.

* Add 4sqlem13m to iset.mm

This is 4sqlem13 from set.mm with not empty changed to inhabited.  The
proof is similar to the set.mm proof with intuitionizing in
various places including using 4sqlemsdc to show decidability.