Add 4sq to iset.mm (#4870)
* Add 4sqlem14 to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with a few changes
related to showing decidability.

* Add 4sqlem15 to iset.mm

Stated as in set.mm.  The proof needs a very small change, for using
apartness when dividing, but is otherwise the set.mm proof.

* Add 4sqlem16 to iset.mm

Stated as in set.mm.  The proof needs to be changed in one place to use
apartness when dividing but is otherwise the set.mm proof.

* Add 4sqlem17 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in a variety of
places but is basically the set.mm proof.

* Add 4sqlem18 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in a few places but
is basically the set.mm proof.

* Add 4sqlem19 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in a few places but
is basically the set.mm proof.

* Copy 4sq from set.mm to iset.mm

Update comments in iset.mm which can now refer to 4sq .

Update mmil.html and mm_100.html