Show pw2en equivalent to excluded middle (#4874)
* Remove redundant necon4bd entry in mmil.html

The more useful one is the one which tells us there is a version which
has a decidability condition.

* Add pw2f1odclem to iset.mm

This is pw2f1olem from set.mm with a decidability condition added.  I'm
not sure the decidability condition is weaker than excluded middle
itself but might as well state it in terms of what this proof needs.

* Add pw2f1odc to iset.mm

This is pw2f1o from set.mm with a decidability condition added (which is
probably almost equivalent to excluded middle, but this is what we
directly need in this proof).

* Add exmidpw2en to iset.mm

This shows that pw2en from set.mm is equivalent to excluded middle.
Update mmil.html and mm_100.html