Add pcadd2 to iset.mm (#4879)
* Rename addid1d to addridd in iset.mm

This is to follow a rename which has already been done in set.mm.

* Add xqltnle to iset.mm

This is similar to theorems for other sets with decidable equality such
as qltnle .

* Add pcxqcl to iset.mm

This is similar to pcxcl but restricts the possible values a bit more.

* Add pcadd2 to iset.mm

Stated as in set.mm.  The proof is similar to the set.mm proof but
changes xrltnle to xqltnle and makes related changes.

Update mmil.html accordingly.

* Remove ltneOLD from mmil.html

It is no longer in set.mm.