Add elfvm to iset.mm
This is a theorem about inhabited function values.  There are somewhat
similar things in set.mm but because it tends to use nonempty sets and
excluded middle in such places, they look somewhat different.

Although I'm not sure whether to say this more explicitly, F doesn't
have to be a function.  It just has to behave like a function around
the point B .