add psrelbas to iset.mm (#4898)
* Add elfvm to iset.mm

This is a theorem about inhabited function values.  There are somewhat
similar things in set.mm but because it tends to use nonempty sets and
excluded middle in such places, they look somewhat different.

Although I'm not sure whether to say this more explicitly, F doesn't
have to be a function.  It just has to behave like a function around
the point B .

* Add elfv2ex to mmil.html

Also, expand the elfvdm entry

* Add basm to iset.mm

This is a relatively straightforward application of elfvm but applies some
extensible structure tweaks to it.

* Add relelbasov to iset.mm

This is elbasov from set.mm with an additional condition.  The proof is
a new one.

* Add fnpsr to iset.mm

The proof is very similar to the proof of psrex .  In fact, replace
psrex with this - psrex would follow as an easy consequence but the more
common practice in set.mm appears to be to have theorems such as
fnmap which we can then apply directly.

* Add psrelbas to iset.mm

Stated as in set.mm.  Although the proof is basically the set.mm one, a
variety of steps need intuitionizing for set existence and using
relelbasov instead of elbasov .

* copy psrelbasfun from set.mm to iset.mm