intuitionize a bit more of mPwSer (#4907)
* Add psrplusgg to iset.mm

This is psrplusg from set.mm with additional set existence conditions.
The proof is similar to the set.mm proof but much longer because of the
need to demonstrate that various classes are sets.

* Add psradd to iset.mm

Stated as in set.mm.  The proof needs rearrangement to use relelbasov
for set existence rather than having it built in to psrplusg .

* Add psraddcl to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in a variety of
places but is basically the set.mm proof.

* add rhmpsrlem1 and rhmpsrlem2 to mmil.html

* add psrmulr and friends to mmil.html

The others are psrmulfval and psrmulval

* add psrmulcllem , psrmulcl to mmil.html

* add psrsca to mmil.html

* add psrvscafval to mmil.html