initial df-plfl work (#4910)
* split definitions

'z' becomes 'c' so it is not confused with the 'z' later in the definition

* fix df-sfl1: m must be univariate

* fix `deg1` syntax

`deg1` is a function not an operation - example: deg1val

1. fix spacing
1. fix `deg1` use in ~ df-plfl

* move rspsnel to main, rename ?spsnel to el?spsn

I proved elrspsn, but it was already proved, so instead move it to main:
https://github.com/metamath/set.mm/issues/4896#issuecomment-2990228826

* Better comment for ~ isdomn5

isdomn5 connects isdomn and isdomn2

isdomn6 moved up since it's similar to isdomn2 and more intuitive, I think.
Originally I thought isdomn5 connected with isdomn6.

* generalize and move r1pid2 to main (+library)

* simplify df-plfl, justified by r1peuqusdeg1 (new) + r1pid2

* fix ci

* rewrap

* wl suggestion

* more explanation for ellcsrspsn

* av suggestion: rename lspsnel5{,a} to lspsnel5{b,}