Revise df-gsum comment in set.mm (#4913)
A and B had not been defined, k was mentioned but that made no sense,
and other minor tweaks to wording.

iset.mm has bigger issues when it comes to df-gsum but for now just try
to keep its comment more or less in sync with the set.mm one.