revise gsum definition in iset.mm (#4924)
* Revise gsum definition in iset.mm

The new definition just requires decidability for whether the index set
is (/) or ( m ... n ) for integers m,n.  It does not try to handle other
cases.

* some comment/web updates for df-igsum

* fix to gsumvalx entry in mmil.html to refer to df-igsum

* Revise df-igsum

Try putting check for dom f = (/) inside the iota.  Hope is that it will
be easier to show set existence for w gsum f .

* Add iotaexel to iset.mm

This is another iotaex replacement.  It is basically a small variation
on riotaexg but expressed somewhat differently.

* Add fngsum to iset.mm

Showing df-igsum is defined everywhere is a bit tricky, but for a given
w and f there are only ZZ X. ZZ (plus one) possible values which feed
into the iota expression, letting us use iotaexab .

Also a minor edit to the header of the Iterated sums in a magma section
(in set.mm and well as iset.mm) - it isn't really necessary to call out
a specific part of the df-gsum (or df-igsum ) comment, but even if there
were "Remark 2" wouldn't work because that comment doesn't have anything
called that.

* Add igsumvalx to iset.mm

Similar in spirit to gsumvalx from set.mm but differs based on our
definition of gsum .  The proof also needs significant changes for set
existence.

* Add igsumval to iset.mm

This is gsumval in set.mm but adjusted for the differences between
gsum definitions.  The proof is similar to the set.mm proof.

* Add gsumpropd to iset.mm

Stated as in set.mm.  The proof needs to be adjusted for differences in
gsum but it is much the same proof as set.mm.  Trim back part of the
comment which I think might be about parts of the set.mm gsum definition
not present in iset.mm.

* Add seqfeq4g to iset.mm

This is seqfeq4 from set.mm with two set existence conditions added.
The proof is by induction and uses seq3clss and some other sequence
theorems which will behave well when everything involved can be shown to
be a set.

* Add gsumpropd2 to iset.mm

Stated as in set.mm.  The proof is similar to a portion of the set.mm
gsumpropd2lem proof but has quite a few differences many of which stem
from the differences between gsumvalx (set.mm) and igsumvalx (iset.mm).

* copy gsummgmpropd from set.mm to iset.mm

* Add gsumress to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in several ways but
is basically the set.mm proof.

* Add gsum0g to iset.mm

This is gsum0 from set.mm with a set existence condition added.  The
proof is a new one but it follows relatively straightforwardly from
igsumval .

Also revise gsumval1 entry in mmil.html.

* Add gsumval2 to iset.mm

Stated as in set.mm.  The proof is a new one, based on igsumval .

* copy uzidd from set.mm to iset.mm

* Add gsumsplit1r to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof but needs
many changes, mostly for differences in seq theorems.

* Add gsumprval to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof, but needs
intuitionizing for differences in seq theorems.

* copy gsumpr12val from set.mm to iset.mm

* Add mulgnngsum to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof but needs
some intuitionizing, especially for differences in seq theorems.

* Add mulgnn0gsum to iset.mm

Stated as in set.mm.  The proof needs a bit of intuitionizing to use
basmex and gsum0g but other than that is the set.mm proof.