Equinumerosity of the natural numbers and ` NN+oo ` is equivalent to the Limited Principle of Omniscience (LPO). (#4931)
* Add propositional logic theorems to mmil.html

This is jad (replacement jaddc ), mt3 , condan (replacement condandc ),
exmid (replacement exmiddc ), pm2.13 (replacement pm2.13dc ),
pm5.62 (replacement pm5.62dc ), pm5.63 (replacement pm5.63dc ),
pm4.83 (replacement pm4.83dc ), and 4cases .

* Add ifbi converse theorems to iset.mm

This is ifnetruedc , ifnefals , and ifnebibdc .

ifnefals is copied from set.mm without change.

ifnetruedc and ifnebibdc are based on ifnetrue and ifnebib from set.mm
with decidability conditions added to the conclusions and to suitable
stages in the proofs.

* Add nninfinf to iset.mm

The mapping of the natural numbers into NN+oo is the obvious one,
which maps n to a sequence of n ones followed by zeroes.  Although it
took me a bit of trickery to prove this mapping is one-to-one, it
isn't especially hard.

* Add xnn0nnen to iset.mm

Since NN0* adds one element to NN0 it is equinumerous to NN0 (and thus
to NN ).  The proof is by showing a one-to-one mapping from NN0*
onto ( ZZ>= ` -u 1 )

* Add nninfninc to iset.mm

This theorem, a consequence of the nonincreasing nature of an element of
NN+oo , is designed to make it easier to work with elements of NN+oo .
The proof is by induction.

* Add nninfct to iset.mm

The limited principle of omniscience (LPO) implies that ` NN+oo ` is
countable.

Includes lemma nninfctlemfo which contains most of the proof.
nninfct takes nninfctlemfo and extends it a bit to use
_om in place of NN0* and express countability using disjoint union
in the usual way.

* Add nnnninfen to iset.mm

This is a passing comment in the cited paper.  Our proof is from
nninfwlpo and ctinf for one direction of the biconditional
and nninfomni for the other.