add a few more continuity theorems to iset.mm (#4935)
* Add cnheibor to mmil.html

* Add addcncf to iset.mm

Stated as in set.mm.  The proof needs small adjustments but is basically
the set.mm proof.

* Add subcncf to iset.mm

Stated as in set.mm.  The proof needs small changes but is basically the
set.mm proof.

* Add mpomulf to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof but needs
a little bit of intuitionizing.

* Add divcncf to mmil.html

* Fix subcncf comment

Change "addition" to "subtraction".  In set.mm and iset.mm.