Add ivthdich to iset.mm
This shows that we will be unable to prove the Intermediate Value
Theorem, by showing that it implies real number dichotomy (a
constructive taboo).

Includes the lemma ivthdichlem and the theorem dich0 (which
provides two equivalent ways to state real number dichotomy).

Includes hoverlt1 and hovergt0 which are two more observations about the
hover function.