The intermediate value theorem implies a constructive taboo (#4943)
* Add ivthreinc to iset.mm

This takes a statement of the intermediate value theorem as a hypothesis
and restates it with some classes (to make it easier to work with).

* Add divcncfap to iset.mm

This is divcncf from set.mm with not equal changed to apart.  The proof
is based on the set.mm proof with the changes being to change not equal
to apart.

* Add maxcncf to iset.mm

This follows readily from maxabs but sometimes it will be more
convenient to use max than express everything in terms of absolute
value.

* Add mincncf to iset.mm

This is like maxcncf but for the minimum of two real numbers.  The proof
is the same as proof of maxcncf but changes addition to subtraction in
the relevant places.

* copy idcncf from set.mm to iset.mm.

* Add sub1cncf to iset.mm

Stated as in set.mm.  The proof needs minor changes but is basically the
set.mm proof.

* Add sub2cncf to iset.mm

Stated as in set.mm.  The proof needs very small changes from the set.mm
proof but is otherwise unchanged.

* Add hovercncf to iset.mm

This states the hover function and shows it is continuous.  The proof is
fairly straightforward from existing theorems about continuity.

* Add hovera to iset.mm

A point at which the hover function is less than a given value.
It is readily provable from existing min/max theorems.

* Add hoverb to iset.mm

Counterpart to hovera for greater than rather than less than, with a
similar proof to hovera .

* Add ivthdich to iset.mm

This shows that we will be unable to prove the Intermediate Value
Theorem, by showing that it implies real number dichotomy (a
constructive taboo).

Includes the lemma ivthdichlem and the theorem dich0 (which
provides two equivalent ways to state real number dichotomy).

Includes hoverlt1 and hovergt0 which are two more observations about the
hover function.

* Wording fix to ivthdich comment.