Add 2idlval to iset.mm
Turns out there is a trick which lets us prove this without the set
existence condition found in 2idlvalg .  Document this trick in
mmil.html (not the first time iset.mm has used it, but worth a
mention as it comes up from time to time and might not be obvious).

Move 2idlmex slightly earlier in the file so we can use it in this
proof.