Add znfi to iset.mm
Stated as in set.mm.  This is a new proof (based on a portion of the
znhash proof which uses fzofig and enfii ), because the set.mm proof
uses hashclb which iset.mm doesn't have and is, I suppose, unlikely to
get.

Put znfi before znhash and use it to shorten the znhash proof.