More Z/nZ theorems in iset.mm (#4962)
* Add residfi to iset.mm

Stated as in set.mm.  The proof needs intuitionizing in various places
but is similar to the set.mm proof.

* copy grpmgmd from set.mm to iset.mm

* Add lgsqrlem4 to mmil.html

Includes a note about how to intuitionize one part of the proof.

* copy invghm from set.mm to iset.mm

* Add ringlghm to iset.mm

Stated as in set.mm.  The proof is the same as the set.mm proof with
some changes for set existence.

* Add ringrghm to iset.mm

Stated as in set.mm.  The proof is the set.mm one with some changes to
use ringcl in place of ovex .

* copy ghmfghm from set.mm to iset.mm

* copy resrhm from set.mm to iset.mm

* copy resrhm2b from set.mm to iset.mm

* copy rhmeql from set.mm to iset.mm

* copy rhmima from set.mm to iset.mm

* copy rnrhmsubrg from set.mm to iset.mm

* Add rhmpropd to iset.mm

Stated as in set.mm.  The proof needs a fair number of changes for set
existence but is otherwise the set.mm proof.

* Add zrhval to iset.mm

Stated as in set.mm.  The proof is fairly different from the set.mm
proof but the needed set existence steps can be added without changing
the conclusion.

* copy zrhpropd from set.mm to iset.mm

* copy znzrh from set.mm to iset.mm

* Add 2idlval to iset.mm

Turns out there is a trick which lets us prove this without the set
existence condition found in 2idlvalg .  Document this trick in
mmil.html (not the first time iset.mm has used it, but worth a
mention as it comes up from time to time and might not be obvious).

Move 2idlmex slightly earlier in the file so we can use it in this
proof.

* Add divsfval to iset.mm

Stated as in set.mm.  The iset.mm proof is a new one.  It does have to
go through a number of steps, perhaps non-obvious ones, but it is
possible to prove the theorem as stated as in set.mm.

* Add qusrhm to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with some small
changes around set existence.

* copy znzrh2 from set.mm to iset.mm

* Add znzrhval to iset.mm

Stated as in set.mm.  The proof is the set.mm proof but is expanded
because a number of additional steps are needed for set existence.

* Add znzrhfo to iset.mm

Stated as in set.mm.  The proof needs several additional steps for set
existence but is otherwise the set.mm proof.

* Rename lspsnel to ellspsn in iset.mm

This rename has already been done in set.mm.

* Add rspsn to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof but
requires adjustment for differences in extensible structure theorems.

* Adjust rlmsca2 entry in mmil.html

Point to the closest replacement we have so far.

* Add zndvds to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with small changes
for set existence.

* copy zndvds0 from set.mm to iset.mm

* copy raleqtrdv and friends from set.mm to iset.mm

This is raleqtrdv , rexeqtrdv , raleqtrrdv , and rexeqtrrdv .

* Add znf1o to iset.mm

Stated as in set.mm.  The proof is the set.mm proof except:
* some changes to express that the proposition in the if
  statement is decidable
* modulus theorems take rationals rather than reals (but in this case it
  is an integer, so this is an easy change)

* Revise zzngim entry in mmil.html

* copy znle2 from set.mm to iset.mm

* Add znleval to iset.mm

Stated as in set.mm.  The proof is the set.mm proof except that it needs
a few changes for set existence.

* copy znleval2 from set.mm to iset.mm

* Revise zntos entry in mmil.html

* Add znhash to iset.mm

Stated as in set.mm.  The proof is the set.mm proof but needs some
intuitionizing in a number of places.

* Small fix to znzrhfo comment

Write ` Z/nZ ` instead of ` ZZ / n ZZ ` .  They look extrememly similar
once rendered to HTML.