Add a few more Z/nZ theorems to iset.mm (#4966)
* Add znfi to iset.mm

Stated as in set.mm.  This is a new proof (based on a portion of the
znhash proof which uses fzofig and enfii ), because the set.mm proof
uses hashclb which iset.mm doesn't have and is, I suppose, unlikely to
get.

Put znfi before znhash and use it to shorten the znhash proof.

* add znfld to mmil.html

* add znidomb to mmil.html

* add znchr to mmil.html

* Add znunit to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with changes in one
portion related to changing dvdsr2 to dvdsr2d .

* Add znunithash to mmil.html

* add znrrg to mmil.html

* add cygzn to mmil.html

* add cygth and cyggic to mmil.html

* add frgpcyg to mmil.html

* add freshmansdream to mmil.html