Add Gauss's Lemma for integer 2 to iset.mm (#4972)
* revise lgsqr entry in mmil.html

* Add lgsqrmod and friends to mmil.html

This is lgsqrmod , lgsqrmodndvds , lgsdchrval and lgsqrmod

* copy gausslemma2dlem0a from set.mm to iset.mm

* copy gausslemma2dlem0b from set.mm to iset.mm

* copy gausslemma2dlem0c from set.mm to iset.mm

* Add gausslemma2dlem0d to iset.mm

Stated as in set.mm.  The proof is basically the set.mm but needs
intuitionizing in places.

* copy gausslemma2dlem0e from set.mm to iset.mm

* copy gausslemma2dlem0f from set.mm to iset.mm

* Add subhalfhalf to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with not equal
changed to apart in one place.

* Add fldiv4lem1div2uz2 to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof but needs
intuitionizing in various simple ways.

* Add fldiv4lem1div2 to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof but needs
intuitionizing in several places.

* copy gausslemma2dlem0g from set.mm to iset.mm

* copy gausslemma2dlem0h from set.mm to iset.mm

* Add gausslemma2dlem0i to iset.mm

Stated as in set.mm.  The proof needs intuitionizing to use closure
theorems in place of ovex , to use the versions of modulus theorems
which take rationals rather than reals, and related changes to use ph as
an antecedent in more steps.

* Add qdclt to iset.mm

This is just like zdclt but for rational numbers instead of integers.
The proof is very similar to the zdclt proof.

* Add gausslemma2dlem1a to iset.mm

Stated as in set.mm.  The proof is basically the set.mm proof, but needs
changes to explicitly state decidability (when doing case elimination on
a decidable proposition) and to change not equal to apart,

* Add rneqdmfinf1o to mmil.html

* Add gausslemma2dlem1 to iset.mm

Stated as in set.mm.  The proof needs a number of changes from the
set.mm proof.

Includes lemma gausslemma2dlem1f1o which replaces what in set.mm is
handled by rneqdmfinf1o .

Includes lemma gausslemma2dlem1cl (a simple closure lemma, but one we
end up using multiple times in the proof of gausslemma2dlem1f1o ).

* copy fvmptd2 from set.mm to iset.mm

* Add gausslemma2dlem2 to iset.mm

Stated as in set.mm.  The proof is very similar to the set.mm
proof but needs one change to use a closure theorem in place of ovex .

* Add gausslemma2dlem3 to iset.mm

Stated as in set.mm.  The proof is the set.mm one with not equal changed
to apart, ovex changed to closure theorems, and rationals rather than
reals used with floor theorems.

* Add gausslemma2dlem4 to iset.mm

Stated as in set.mm.  The proof needs intuitionizing especially for
rational numbers in floor theorems and set existence.  Some of the logic
is also rearranged, mostly to avoid a use of excluded middle, but while
I'm at it, to simply some tangled steps.

* Add gausslemma2dlem5a to iset.mm

The set.mm proof needs intuitionizing around using fzfigd and rational
numbers for modulus theorems.  This also leads to some rearrangement and
added steps, but it is basically the set.mm proof.

* Add gausslemma2dlem5 to iset.mm

Stated as in set.mm.  The proof needs changes to use fzfigd and to use
rational numbers with floor theorems, but is otherwise the set.mm proof.

* Add gausslemma2dlem6 to iset.mm

Stated as in set.mm.  The proof needs changes to use fzfigd and to use
rational numbers in modulus/floor theorems, but is otherwise the set.mm
proof.

* Add gausslemma2dlem7 to iset.mm

Stated as in set.mm.  The proof needs changes in one portion to use
rational numbers in modulus theorems but is otherwise unchanged from the
set.mm proof.

* Add gausslemma2d to iset.mm

Stated as in set.mm.  The proof is the set.mm proof except that a few
parts are changed to use rational numbers for modulus theorems.