Add RLReg , Domn , and IDomn to iset.mm (#4974)
* Add RLReg , Domn , and IDomn to iset.mm

Copied without change from set.mm.  Consists of the syntaxes, df-rlreg ,
df-domn , and df-idom

* Add rrgmex to iset.mm

This is a set existence theorem for RLReg , similar to basmex ,
lssmex , etc.

* Add rrgval to iset.mm

Stated as in set.mm.  The proof is rearranged significantly to change
the way we do set existence, but it is based on the set.mm proof.

* copy isrrg from set.mm to iset.mm

* copy rrgeq0i from set.mm to iset.mm

* copy rrgeq0 from set.mm to iset.mm

* add rrgsupp to mmil.html

* copy rrgss from set.mm to iset.mm

* Add unitrrg to iset.mm

Stated as in set.mm.  The proof is the set.mm proof with some small
changes to use unitcld in place of unitcl .

* copy ringlzd and ringrzd from set.mm to iset.mm

* copy rrgnz from set.mm to iset.mm

* Add isdomn to iset.mm

Stated as in set.mm.  The proof needs a few changes for set existence
but is otherwise the set.mm proof.

* copy domnnzr from set.mm to iset.mm

* copy domnring from set.mm to iset.mm

* copy domneq0 from set.mm to iset.mm

* copy domnmuln0 from set.mm to iset.mm

* add isdomn5 to mmil.html

* add isdomn2 to mmil.html

* add domnrrg to mmil.html

* add isdomn6 to mmil.html

* add isdomn3 to mmil.html

* add isdomn4 to mmil.html

* Add isnzr2 to iset.mm

Stated as in set.mm.  The proof is the same as the set.mm proof except
for one section which is extensively rewritten to use 2o ~<_ B rather
than 1o ~< B (but which works much the same way).

* Add isnzr2hash to mmil.html

* add nzrpropd to mmil.html

* Add opprnzr to iset.mm

Stated as in set.mm.  The proof is from opprnzrbg .

* Add opprdomnbg to iset.mm

This is opprdomnb from set.mm with an added set existence condition.
The proof is similar to the set.mm proof but with set existence
conditions added throughout.

* Add opprdomn to iset.mm

Stated as in set.mm.  The proof is from opprdomnbg .

* add isdomn4r to mmil.html

* add domnlcanb to mmil.html

* add domnlcan to mmil.html

* add domnrcanb to mmil.html

* add domnrcan to mmil.html

* add domneq0r to mmil.html

* copy isidom from set.mm to iset.mm

* copy idomdomd from set.mm to iset.mm

* Add idomcringd to iset.mm

Copied from set.mm with the only change being to the comment, to remove
a reference to a theorem that iset.mm doesn't have.

* copy idomringd from set.mm to iset.mm