Add lencl to iset.mm
Stated as in set.mm.  The proof is different from the set.mm proof
because the set.mm proof uses finite set theorrems not available
in iset.mm.

Also put it slightly earlier in the file as it will be needed in places
that set.mm does not need it.